move resources around
Summary:

Test Plan:
